+++ constants/meta/Scroll-meta-data/scroll-posY-meta.ts
@@ -1,0 +1,5 @@
+import { Meta } from "../json-schema-meta.ts";
+
+export const scrollPositionYMeta: Meta = {
+    description: "Vertical position to scroll to on the webpage.",
+};
