+++ constants/meta/Scroll-meta-data/scroll-postion-meta.ts
@@ -1,0 +1,6 @@
+
+import { Meta } from "../json-schema-meta.ts";
+export const scrollPositionMeta: Meta = {
+    description:
+        "Coordinates to scroll to on the webpage. Specify `x` and `y` values representing the horizontal and vertical positions respectively.",
+};
