+++ constants/meta/Scroll-meta-data/scroll-postionX-meta.ts
@@ -1,0 +1,6 @@
+
+import { Meta } from "../json-schema-meta.ts";
+
+export const scrollPositionXMeta: Meta = {
+    description: "Horizontal position to scroll to on the webpage.",
+};
