+++ constants/meta/input/input-meta.ts
@@ -1,0 +1,8 @@
+
+import { Meta } from "../json-schema-meta.ts";
+
+export const inputMeta: Meta = {
+    title: "Input Action",
+    description:
+        "Defines an input action where a value is entered into a specified input field identified by the `selector`, `label`, `testID`, `placeholder` or `role` .",
+};
